Introduce message handler trait
